refactor: using enumerate
while effectively the same, this uses enumerate instead of a
range(len(x)) approach.

for more on the differences, this stack overflow was helpful:
https://stackoverflow.com/questions/24150762/pythonic-range-vs-enumerate-in-python-for-loop

ultimately though, in this case there was no benefit since we _knew_ the
input would be a string other tha perhaps a slight performance benefit
from the way enumerate works.